Replace custom decoder with custom PDU
