+++ pyproject.toml
@@ -46,7 +46,7 @@ pre-commit = {version = "^3.8.0", optional = true}
 toml = {version = "^0.10.2", optional = true}
 bump2version = {version = "^1.0.1", optional = true}
 plum-py = "^0.8.0"
-pymodbus = "3.7.4"
+pymodbus = "~3.7.0"
 
 [tool.poetry.extras]
 test = [
